Resampling should always be done on plot range not data range
Using data range is broken in cases where a graph function returns values
way outside of plot range. This leads to undersampling while it shouldn't.

Example:

[++++++++++++++data range++++++++++++++++]
            [plot range]
the sampling algorithm would compare the total number of points in data rage
with the number of pixels in plot range, this is wrong.

The main drawback is that auto-scale is broken now. One option would be to keep
the first and last points after resampling.

Signed-off-by: Alexis Jeandet <alexis.jeandet@member.fsf.org>